token/requirements.txt
boto3_type_annotations
pyjwt
sqlalchemy